backend/requirements.txt
flask==3.1.0
flask-cors>=5.0.0
gunicorn==23.0.0
pydantic>=2.8.0,<3.0.0
python-dotenv==1.1.0
supabase>=2.0.0
google-genai>=1.0.0
resend>=2.0.0
Jinja2>=3.1.0
sentry-sdk[flask]>=2.0.0
fpdf2>=2.8.0
